feat: add cookie-parser middleware, restructure API routes, and create user and post routers
