+++ .editorconfig
@@ -1,0 +1,9 @@
+root = true
+
+[*]
+end_of_line = crlf
+insert_final_newline = false
+indent_style = tab
+charset = utf-8
+trim_trailing_whitespace = true
+tab_width = 4
