+++ app.js
@@ -1,18 +1,20 @@
 const express = require('express');
 const dotenv = require('dotenv')
+const cookieParser = require('cookie-parser');
 const history = require('connect-history-api-fallback');
 const cors = require('cors');
 const app = express();
 
 dotenv.config();
 app
-.use(history())
 .use(cors())
 .use(express.json({limit: '50mb'}))
-.use(express.urlencoded({ extended: true }));
+.use(express.urlencoded({ extended: true }))
+.use(cookieParser());
 
-const apiRouter = require("./src/routes/api.js");
-app.use("/api", apiRouter);
-app.use(express.static('client/dist'));
+const apiRouter = require("./src/routes/index.js");
+app.use("/api", apiRouter)
+.use(history())
+.use(express.static('client/dist'));
 
 module.exports = app;
