+++ package.json
@@ -18,6 +18,7 @@
     "bcrypt": "^5.1.1",
     "concurrently": "^9.1.2",
     "connect-history-api-fallback": "^2.0.0",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
